requirements-test.txt
pytest
requests  # workaround for https://forge.softwareheritage.org/T2634
requests-mock
